app1/requirements.txt
uvicorn
fastapi
requests
python-consul